+++ include/circt/LogicalEquivalence/Circuit.h
