+++ include/circt/LogicalEquivalence/Solver.h
