+++ include/circt/LogicalEquivalence/Utility.h
