+++ lib/CMakeLists.txt
@@ -7,9 +7,9 @@ add_subdirectory(Bindings)
 add_subdirectory(CAPI)
 add_subdirectory(Conversion)
 add_subdirectory(Dialect)
+add_subdirectory(LogicalEquivalence)
 add_subdirectory(Reduce)
 add_subdirectory(Scheduling)
 add_subdirectory(Support)
 add_subdirectory(Target)
 add_subdirectory(Transforms)
-add_subdirectory(Verification)
