+++ lib/LogicalEquivalence/Solver.cpp
@@ -10,10 +10,10 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#include "circt/Verification/Solver.h"
-#include "circt/Verification/Circuit.h"
-#include "circt/Verification/LogicExporter.h"
-#include "circt/Verification/Utility.h"
+#include "circt/LogicalEquivalence/Solver.h"
+#include "circt/LogicalEquivalence/Circuit.h"
+#include "circt/LogicalEquivalence/LogicExporter.h"
+#include "circt/LogicalEquivalence/Utility.h"
 #include "mlir/IR/Builders.h"
 #include <string>
 #include <z3++.h>
